fix(nx): fix gather C implementation
This is a temporary commit just to save and tidy up (e.g. removing
the debug printing and manual macro expansion etc.) later.

caml_nx_gather_bc was relying on the bigarray dimensions (ba_z->dim) of
the result (v_z) array to figure out the shape of the index array.

I believe this is incorrect because that field infact states the
_physical_ shape of the bigarray, rather than its logical shape.

This resulted in a situation where, indices_shape.(i) > data_shape.(i)
for a non-axis dimension in the C code, despite this being checked for
in the OCaml code.

For example, with the below test:

    let test_slice_with_steps ctx () =
      let t =
        Nx.create ctx Nx_core.Dtype.float32 [| 3; 4 |]
          (Array.init 12 float_of_int)
      in
      let s = Nx.slice_ranges ~steps:[ 2; 2 ] [ 0; 0 ] [ 3; 4 ] t in
      check_t "slice with steps" [| 2; 2 |] [| 0.; 2.; 8.; 10. |] s

We had the following (truncated, relevant) debug printing:
    dim 0: data(logical=3 ba=12)  indices=8  z=8
    dim 1: data(logical=4 ba=1024)  indices=1024  z=1024
    nx_c_gather_float: shape mismatch at dim 1 (indices=1024 > data=4)

The solution is to extend the C function to take the logical shape of
the indices as an argument, using it for both the index array and the
output array.