feat(nx): Refactor einsum
This commit refactors the einsum implementation to fix a bug in the
previous version.

The way it previously calculated axes for contraction/tensordot
was incorrect, because it calculated _all_ contracted axes for a given
input, leading to a different number of axes for each input, rather
than the ones in common (which by definition, have the same number).

It still fails the test for transpose, and writing this, I'm also
suspicious of how I'm handling an n-ary input fold, specifically how the
axes mapping and the index_var sets are updated on the accumulating
parameter.