feat(nx): Refactor einsum again
The previous implementation of n-ary einsum was incorrect because the
change in index vars and dimensions as the outer products/contractions
proceeded was incorrect.

This commit fixes this with several changes. First, binary_einsum
takes two shapes and two index var strings, checks the dimensions,
and the computes their respective contraction axes, resulting shape,
and resulting index variables.

This information is then use to construct a tree (einsum_plan) of outer
products and contractions. The plan can be used to check the final
index vars string is a permutation of the output string.

This tree (einsum_plan) is evaluated, leaving the output modulo any
transpositions.

Finally, the evaluated index var string is used to construct the
transposition axes.

This version passes the current einsum tests, and thus generalises
matrix multiplication and transposition.